refactor: move pool-stake components
