+++ packages/web/src/views/pool/pool-stake/components/select-liquidity-list/SelectLiquidityList.stories.tsx
