+++ packages/web/src/views/pool/pool-stake/components/select-liquidity-list/select-lilquidity-list-item/SelectLiquidityListItem.stories.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import SelectLiquidityListItem from "./SelectLiquidityListItem";
-import { STAKED_OPTION } from "@constants/option.constant";
 import { action } from "@storybook/addon-actions";
 
 export default {
