+++ packages/web/src/views/pool/pool-stake/components/select-liquidity/SelectLiquidity.styles.ts
