+++ packages/web/src/views/pool/pool-stake/components/select-stake-result/SelectStakeResult.stories.tsx
