+++ packages/web/src/views/pool/pool-stake/components/stake-position-modal/StakePositionModal.styles.ts
