+++ packages/web/src/views/pool/pool-stake/components/stake-position/StakePosition.styles.ts
